chore(tests): rename folder
