+++ tests/Serilog.Ui.Web.Tests/Authorization/AuthorizationFilterTest.cs
@@ -2,7 +2,7 @@
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Ui.Web.Tests.Util;
+using Ui.Web.Tests.Utilities;
 using Xunit;
 
 namespace Serilog.Ui.Web.Tests.Authorization;
